\name{HDCytoData}
\docType{data}
\alias{HDCytoData}
\alias{HDCytoData-package}
\title{High-dimensional cytometry data sets in 'SummarizedExperiment' and 'flowSet' formats}
\description{
A set of publicly available high-dimensional cytometry data sets, formatted into 'SummarizedExperiment' 
and 'flowSet' objects, containing expression values as well as meta-data including sample IDs, 
group IDs, population labels, and protein marker names.
}
\details{
This package contains a set of publicly available high-dimensional cytometry data sets, formatted 
into 'SummarizedExperiment' and 'flowSet' objects. The objects contain expression values as 
well as meta-data including sample IDs, group IDs (where relevant), population labels, and 
protein marker names.

These data sets have been used in our previous publications for benchmarking evaluations. 
They are provided here as 'SummarizedExperiments' and 'flowSets' to make them easier to access 
for other method developers.

For details on each data set, see the individual help files for each data set.

The steps to prepare each data set are included in the script \code{inst/scripts/make-data.R}
}
\keyword{datasets}
